// apps/api/src/course/enums/task-name.enum.ts
export enum TaskName {
  CREATE_COURSE = 'CREATE_COURSE',
  GENERATE_TITLE = 'GENERATE_TITLE',
  SET_AUDIENCE = 'SET_AUDIENCE',
  SET_OBJECTIVE = 'SET_OBJECTIVE',
  SET_BUILDING_METHOD = 'SET_BUILDING_METHOD',
  SET_MODULES = 'SET_MODULES',
  SET_UNITS = 'SET_UNITS',
  GET_EXERCISE_TYPES = 'GET_EXERCISE_TYPES',
  SET_EVALUATION = 'SET_EVALUATION',
  SET_EVALUATION_DETAILS = 'SET_EVALUATION_DETAILS',
  SET_BRANDING = 'SET_BRANDING',
  GENERATE_COURSE = 'GENERATE_COURSE',
}